+++ app/js/preloads/actions.js
@@ -215,13 +215,3 @@ ipcRenderer.on("search-element-click", () => {
         focussedSearchElement.click()
     }
 })
-
-window.addEventListener("keydown", e => {
-    if (e.code === "F11") {
-        e.preventDefault()
-    }
-})
-
-window.addEventListener("mousemove", e => {
-    ipcRenderer.sendToHost("top-of-page-with-mouse", !e.clientY)
-})